Change data.json to be contribution level
This simplifies the data structure of the data.json file
and might make it easier to do other types of analysis.

It moves the aggregation logic down to the route level but
that seems fine because it's simply logic and changes the
need for understanding a nested data model to the routes
that need it.